Got it linking and sending a message to Discord
Currently when it tries to connect to the gateway, it gets this error returned: ```Unable to connect to gateway.discord.gg on port 443, error: error in handshake : X509 - Certificate verification failed, e.g. CRL, CA or signature check failed``` but I did get it sending a message to discord's servers to receive the gateway url. Thats about it...