Remove old event system, event file changes, and dispatch events
I removed the old event system so there is no more traces of that. I had to prefix every discordpp include in the event files with "../", there would be errors when the user would try to use a event. I also now dispatch all events with the new event system.